pyproject.toml
[project]
name = "chatbot-with-chainlit-and-openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.13",
]
